+++ ledger/api/character/planetary.py
@@ -4,9 +4,12 @@ from ninja import NinjaAPI
 
 from django.db.models import Q
 
+from allianceauth.authentication.models import UserProfile
+
 from ledger.api import schema
 from ledger.api.helpers import get_alts_queryset, get_character
 from ledger.hooks import get_extension_logger
+from ledger.models.characteraudit import CharacterAudit
 from ledger.models.planetary import CharacterPlanet, CharacterPlanetDetails
 
 logger = get_extension_logger(__name__)
@@ -63,12 +66,13 @@ class LedgerPlanetaryApiEndpoints:
             tags=self.tags,
         )
         def get_planetarydetails(request, character_id: int, planet_id: int):
+            request_main = request.GET.get("main", False)
             perm, main = get_character(request, character_id)
 
             if not perm:
                 return 403, "Permission Denied"
 
-            if character_id == 0:
+            if character_id == 0 or request_main:
                 characters = get_alts_queryset(main)
             else:
                 characters = [main]
@@ -109,4 +113,40 @@ class LedgerPlanetaryApiEndpoints:
                         "extractors": extractors,
                     }
                 )
+            return output
+
+        @api.get(
+            "account/planetary/admin/",
+            response={200: List[schema.CharacterAdmin], 403: str},
+            tags=self.tags,
+        )
+        def get_character_admin(request):
+            chars_ids = CharacterAudit.objects.visible_eve_characters(
+                request.user
+            ).values_list("character_id", flat=True)
+
+            users_char_ids = UserProfile.objects.filter(
+                main_character__character_id__in=chars_ids
+            )
+
+            if not chars_ids:
+                return 403, "Permission Denied"
+
+            character_dict = {}
+
+            for character in users_char_ids:
+                # pylint: disable=broad-exception-caught
+                try:
+                    character_dict[character.main_character.character_id] = {
+                        "character_id": character.main_character.character_id,
+                        "character_name": character.main_character.character_name,
+                        "corporation_id": character.main_character.corporation_id,
+                        "corporation_name": character.main_character.corporation_name,
+                    }
+                except Exception:
+                    continue
+
+            output = []
+            output.append({"character": character_dict})
+
             return output