+++ ledger/urls.py
@@ -5,7 +5,11 @@ from django.urls import path, re_path
 from ledger.api import api
 from ledger.views.character.char_audit import add_char, fetch_memberaudit
 from ledger.views.character.character_ledger import character_admin, character_ledger
-from ledger.views.character.planetary import planetary_ledger, switch_alarm
+from ledger.views.character.planetary import (
+    planetary_admin,
+    planetary_ledger,
+    switch_alarm,
+)
 from ledger.views.corporation.corp_audit import add_corp
 from ledger.views.corporation.corp_events import (
     create_event,
@@ -58,12 +62,17 @@ urlpatterns = [
     path("events/<int:event_id>/delete/", delete_event, name="delete_event"),
     path("events/ajax/load_events", load_events, name="load_events"),
     # -- -- Planetary
-    path("planetary/<int:character_pk>/", planetary_ledger, name="planetary_ledger"),
+    path(
+        "planetary_ledger/<int:character_pk>/",
+        planetary_ledger,
+        name="planetary_ledger",
+    ),
     path(
         "planetary/switch_alarm/<int:character_id>/planet/<int:planet_id>/",
         switch_alarm,
         name="switch_alarm",
     ),
+    path("planetary_admin/", planetary_admin, name="planetary_admin"),
     # -- API System
     re_path(r"^api/", api.urls),
 ]